+++ News ToDay/Custom Views/TabBarControllers/NTDTabBarController.swift
@@ -9,6 +9,18 @@ import UIKit
 
 class NTDTabBarController: UITabBarController {
     
+    var router: RouterProtocol
+    
+    init(router: RouterProtocol) {
+        self.router = router
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -19,7 +31,7 @@ class NTDTabBarController: UITabBarController {
     }
     
     
-    func createHomeNC() -> UINavigationController {
+    func createHomeNC() -> UIViewController {
         let homeVC        = HomeViewController()
         homeVC.title      = ScreenTitleStrings.browse
         homeVC.tabBarItem = UITabBarItem(title: nil, image: TabBarImage.home, selectedImage: TabBarSelectedImage.home)
@@ -45,9 +57,15 @@ class NTDTabBarController: UITabBarController {
     
     func createProfileNC() -> UINavigationController {
         let profileVC         = ProfileViewController()
+        profileVC.router      = self.router
         profileVC.title       = ScreenTitleStrings.profile
         profileVC.tabBarItem  = UITabBarItem(title: nil, image: TabBarImage.user, selectedImage: TabBarSelectedImage.user)
         return UINavigationController(rootViewController: profileVC)
+    }
+    
+    
+    deinit {
+        print("deinit from NTDTabBarController")
     }
 }
