+++ News ToDay/Models/OnboardingModel.swift
@@ -27,5 +27,5 @@ enum OnboardingMockData {
 //MARK: - Model
 
 struct OnboardingModel {
-    let mainImage: UIImage
+    let mainImage: UIImage?
 }
