+++ News ToDay/Screens/Login+Register/SignUpViewController.swift
@@ -1,5 +1,5 @@
 //
-//  SignInViewController.swift
+//  SignUpViewController.swift
 //  News ToDay
 //
 //  Created by Дарья Большакова on 27.03.2024.
@@ -7,18 +7,25 @@
 
 import UIKit
 
-class SignInViewController: UIViewController, UITextFieldDelegate {
+class SignUpViewController: UIViewController, UITextFieldDelegate {
+    
+    var router: RouterProtocol?
+    
     var passwordIsSecure = true
     let logoLabel = UILabel()
     let textLabel = UILabel()
-    var loginText = SignInStrings.email
-    var passwordText = SignInStrings.password
-    var loginEnterText = SignInStrings.enterEmail
-    var passwordEnterText = SignInStrings.enterPassword
+    var nameText = SignUpStrings.name
+    var loginText = SignUpStrings.email
+    var passwordText = SignUpStrings.password
+    var repeatPasswordText = SignUpStrings.repeatPassword
+    var nameEnterText = SignUpStrings.enterName
+    var loginEnterText = SignUpStrings.enterEmail
+    var passwordEnterText = SignUpStrings.enterPassword
+    var repeatPasswordEnterText = SignUpStrings.enterRepeatPassword
     
     lazy var button: UIButton = {
         let button = UIButton(primaryAction: buttonTapped())
-        button.setTitle(SignInStrings.signIn, for: .normal)
+        button.setTitle(SignUpStrings.signIn, for: .normal)
         button.setTitleColor(.textOnActiveButtonColor, for: .normal)
         button.titleLabel?.font = UIFont(name: "Inter-SemiBold", size: 20)
         button.layer.cornerRadius = 15
@@ -29,7 +36,7 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
     
     lazy var questionLabel: UILabel = {
         let label = UILabel()
-        label.text = SignInStrings.label
+        label.text = SignUpStrings.label
         label.textColor = .textSecondaryColor
         label.heightAnchor.constraint(equalToConstant: 15).isActive = true
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -37,31 +44,31 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
         return label
     }()
     
-    lazy var signUpButton: UIButton = {
-        let button = UIButton(primaryAction: signUpAction())
-        button.setTitle(SignInStrings.signUp, for: .normal)
+    lazy var signInButton: UIButton = {
+        let button = UIButton(primaryAction: signInAction())
+        button.setTitle(SignUpStrings.signIn, for: .normal)
         button.setTitleColor(.textPrimaryColor, for: .normal)
         button.heightAnchor.constraint(equalToConstant: 15).isActive = true
         button.translatesAutoresizingMaskIntoConstraints = false
-        
+
         return button
     }()
     
     lazy var stack: UIStackView = {
         let stack = UIStackView()
-       
+
         stack.axis = .horizontal
         stack.spacing = 3
         stack.alignment = .center
         stack.distribution = .fillProportionally
         stack.translatesAutoresizingMaskIntoConstraints = false
-        
+
         return stack
     }()
     
-    lazy var eye: UIView = {
-        let image = createImageForField(name: "EyeIcon")
-        return button
+    lazy var userIcon: UIView = {
+        let image = createImageForField(name: "UserIcon")
+        return image
     }()
     
     lazy var  envelope: UIView = {
@@ -72,6 +79,17 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
     lazy var lock: UIView = {
         let image = createImageForField(name: "LockIcon")
         return image
+    }()
+    
+    lazy var lock2: UIView = {
+        let image = createImageForField(name: "LockIcon")
+        return image
+    }()
+    
+    lazy var userName: UITextField =  {
+        let field = createTextField(name: nameText, imName: userIcon)
+        field.keyboardType = .emailAddress
+        return field
     }()
     
     lazy var login: UITextField =  {
@@ -86,35 +104,52 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
         return field
     }()
     
+    lazy var repeatPassword: UITextField =  {
+        let field = createTextField(name: repeatPasswordText, imName: lock2)
+        field.isSecureTextEntry = true
+        return field
+    }()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         self.navigationController?.setNavigationBarHidden(true, animated: true)
         setupUI()
         view.backgroundColor = .white
+        
+        userName.delegate = self
+        userName.tag = 0
         login.delegate = self
-        login.tag = 0
+        login.tag = 1
         password.delegate = self
-        password.tag = 1
+        password.tag = 2
+        repeatPassword.delegate = self
+        repeatPassword.tag = 3
+        
         
         button.isUserInteractionEnabled = false
         button.alpha = 0.5
-        
+        createDismissKeyboardTapGesture()
     }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         updateLabelsText()
-        
+    }
+    
+    
+    private func createDismissKeyboardTapGesture() {
+        let tap = UITapGestureRecognizer(target: view, action: #selector(UIView.endEditing(_:)))
+        view.addGestureRecognizer(tap)
     }
     
     
     func setupUI() {
-        logoLabel.text = SignInStrings.title
+        logoLabel.text = SignUpStrings.title
         logoLabel.font = UIFont(name: "Inter-SemiBold" , size: 25)
         logoLabel.textColor = .textPrimaryColor
         logoLabel.textAlignment = .left
         
-        textLabel.text = SignInStrings.text
+        textLabel.text = SignUpStrings.text
         textLabel.font = UIFont(name: "Inter-Regular" , size: 18)
         textLabel.textColor = .textSecondaryColor
         textLabel.textAlignment = .left
@@ -125,14 +160,16 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
         button.translatesAutoresizingMaskIntoConstraints = false
         
         stack.addArrangedSubview(questionLabel)
-        stack.addArrangedSubview(signUpButton)
+        stack.addArrangedSubview(signInButton)
         
-        view.addSubview(logoLabel)
+        view.addSubviews(logoLabel)
         view.addSubviews(textLabel)
+        view.addSubviews(userName)
         view.addSubviews(login)
         view.addSubviews(password)
+        view.addSubviews(repeatPassword)
         view.addSubviews(button)
-        view.addSubview(stack)
+        view.addSubviews(stack)
         
         NSLayoutConstraint.activate([
             logoLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 10),
@@ -143,7 +180,11 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
             textLabel.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
             textLabel.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
             
-            login.topAnchor.constraint(equalTo: textLabel.bottomAnchor, constant: 20),
+            userName.topAnchor.constraint(equalTo: textLabel.bottomAnchor, constant: 20),
+            userName.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
+            userName.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
+            
+            login.topAnchor.constraint(equalTo: userName.bottomAnchor, constant: 20),
             login.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
             login.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
             
@@ -151,12 +192,17 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
             password.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
             password.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
             
-            button.topAnchor.constraint(equalTo: password.bottomAnchor, constant: 60),
+            repeatPassword.topAnchor.constraint(equalTo: password.bottomAnchor, constant: 20),
+            repeatPassword.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
+            repeatPassword.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
+            
+            button.topAnchor.constraint(equalTo: repeatPassword.bottomAnchor, constant: 30),
             button.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
             button.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
             
             stack.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -20),
             stack.centerXAnchor.constraint(equalTo: view.safeAreaLayoutGuide.centerXAnchor, constant: 0)
+            
         ])
     }
     
@@ -186,7 +232,6 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
         view.addSubview(imageView)
         return view
     }
-    
 
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
           if let nextField = textField.superview?.viewWithTag(textField.tag + 1) as? UITextField {
@@ -199,36 +244,28 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
     
     func textFieldDidBeginEditing(_ textField: UITextField) {
         textField.backgroundColor = .white
-        if textField == login {
+        switch textField {
+        case userName:
+            textField.leftView = createImageForField(name: "UserSelectedIcon")
+        case login:
             textField.leftView = createImageForField(name: "EmailSelectedIcon")
-        } else {
+        case password:
             textField.leftView = createImageForField(name: "LockSelectedIcon")
+        case repeatPassword:
+            textField.leftView = createImageForField(name: "LockSelectedIcon")
+        default:
+            break
         }
         textField.layer.borderWidth = 1
         textField.layer.borderColor = UIColor.buttonActiveColor.cgColor
-        if textField == password {
-            
-            textField.rightView = createImageForField(name: "EyeIcon")
-            textField.rightViewMode = .always
-            
-            let eyeBtn = UIButton(primaryAction: eyeAction())
-            eyeBtn.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(eyeBtn)
-            eyeBtn.heightAnchor.constraint(equalToConstant: 30).isActive = true
-            eyeBtn.widthAnchor.constraint(equalToConstant: 30).isActive = true
-            eyeBtn.topAnchor.constraint(equalTo: password.topAnchor, constant: 10).isActive = true
-            eyeBtn.trailingAnchor.constraint(equalTo: password.trailingAnchor, constant: -10).isActive = true
-            
-        }
+
+    }
+    func textFieldDidEndEditing(_ textField: UITextField, reason: UITextField.DidEndEditingReason) {
+        check()
     }
 
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        check()
-        return true
-    }
-    
     func check()  {
-        if self.login.text!.trimmingCharacters(in: .whitespaces) != "" && self.password.text!.trimmingCharacters(in: .whitespaces) != "" {
+        if self.login.text!.trimmingCharacters(in: .whitespaces) != "" && self.password.text!.trimmingCharacters(in: .whitespaces) != "" && self.userName.text!.trimmingCharacters(in: .whitespaces) != "" && self.repeatPassword.text!.trimmingCharacters(in: .whitespaces) != "" {
             self.button.isUserInteractionEnabled = true
             self.button.alpha = 1.0
         }
@@ -237,73 +274,61 @@ class SignInViewController: UIViewController, UITextFieldDelegate {
     private func updateLabelsText() {
         self.navigationController?.navigationBar.isHidden = true
         self.tabBarItem.title = nil
-        logoLabel.text = SignInStrings.title
-        textLabel.text = SignInStrings.text
-        loginText = SignInStrings.email
-        passwordText = SignInStrings.password
-        loginEnterText = SignInStrings.enterEmail
-        passwordEnterText = SignInStrings.enterPassword
-        button.setTitle(SignInStrings.signUp, for: .normal)
-        signUpButton.setTitle(SignInStrings.signIn, for: .normal)
+        logoLabel.text = SignUpStrings.title
+        textLabel.text = SignUpStrings.text
+        
+        nameText = SignUpStrings.name
+        loginText = SignUpStrings.email
+        passwordText = SignUpStrings.password
+        repeatPasswordText = SignUpStrings.repeatPassword
+        
+        nameEnterText = SignUpStrings.enterName
+        loginEnterText = SignUpStrings.enterEmail
+        passwordEnterText = SignUpStrings.enterPassword
+        repeatPasswordEnterText = SignUpStrings.enterRepeatPassword
+        
+        button.setTitle(SignUpStrings.signUp, for: .normal)
+//        signInButton.setTitle(SignUpStrings.signUp, for: .normal)
         
     }
     
     func buttonTapped() -> UIAction {
-        let act = UIAction { _ in
-            if self.login.text!.trimmingCharacters(in: .whitespaces) != "" && self.password.text!.trimmingCharacters(in: .whitespaces) != "" {
-                print("sign in")
-            }
+        let act = UIAction { [weak self] _ in
+            guard let self = self else { return }
             if self.login.text!.trimmingCharacters(in: .whitespaces) == "" {
-                self.login.attributedPlaceholder = NSAttributedString(string: SignInStrings.enterEmail, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+                self.login.attributedPlaceholder = NSAttributedString(string: SignUpStrings.enterEmail, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
                 self.login.layer.borderColor = UIColor.red.cgColor
             }
             if self.password.text!.trimmingCharacters(in: .whitespaces) == "" {
-                self.password.attributedPlaceholder = NSAttributedString(string: SignInStrings.enterPassword, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+                self.password.attributedPlaceholder = NSAttributedString(string: SignUpStrings.enterPassword, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
                 self.password.layer.borderColor = UIColor.red.cgColor
             }
-        }
-        return act
-    }
-    
-    func eyeAction() -> UIAction {
-        let act = UIAction { _ in
-            print("eye")
-            if self.passwordIsSecure {
-                self.passwordIsSecure = false
-                self.password.rightView = self.createImageForField(name: "EyeSelectedIcon")
-                self.password.isSecureTextEntry = false
-            } else {
-                self.passwordIsSecure = true
-                self.password.rightView = self.createImageForField(name: "EyeIcon")
-                self.password.isSecureTextEntry = true
+            if self.userName.text!.trimmingCharacters(in: .whitespaces) == "" {
+                self.userName.attributedPlaceholder = NSAttributedString(string: SignUpStrings.enterName, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+                self.userName.layer.borderColor = UIColor.red.cgColor
+            }
+            if self.repeatPassword.text!.trimmingCharacters(in: .whitespaces) == "" {
+                self.repeatPassword.attributedPlaceholder = NSAttributedString(string: SignUpStrings.enterRepeatPassword, attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+                self.repeatPassword.layer.borderColor = UIColor.red.cgColor
+            }
+            if self.password.text == self.repeatPassword.text {
+                print("sign up")
+            } else{
+                print("nfn")
             }
         }
         return act
     }
     
-    func signUpAction() -> UIAction {
-        let act = UIAction { _ in
-            let vc = SignUpViewController()
-            self.navigationController?.pushViewController(vc, animated: true)
+    func signInAction() -> UIAction {
+        let act = UIAction { [weak self] _ in
+            self?.router?.authFlow()
         }
         return act
     }
     
+    
+    deinit {
+        print("deinit from SignUpViewController")
+    }
 }
-
-
-//import SwiftUI
-//
-//struct Landuage: PreviewProvider {
-//    static var previews: some View {
-//        ContainerView().ignoresSafeArea()
-//    }
-//    struct ContainerView: UIViewRepresentable {
-//        let view = SignInViewController()
-//
-//        func makeUIView(context: Context) -> some UIView {
-//            return view
-//        }
-//        func updateUIView(_ uiView: UIViewType, context: Context) { }
-//    }
-//}
